Modify `Size2` and `Point2` to operate on real numbers rather than integers
It would probably be better to use two separate sets of `Size*` and `Point*` classes for real and whole numbers, but for now it's enough to just have the one with reals.

+++ src/dx/DxSwapChain.cpp
@@ -42,8 +42,8 @@ namespace reni::rhi::dx {
 
 		m_swapChain = swapChain;
 		setBufferSize({
-			static_cast<int>(scDesc.Width),
-			static_cast<int>(scDesc.Height)
+			static_cast<float>(scDesc.Width),
+			static_cast<float>(scDesc.Height)
 		});
 	}
